+++ src/blogmore/markdown/extensions.py
@@ -9,6 +9,7 @@ from typing import Any
 from blogmore.markdown.admonitions import AdmonitionsExtension
 from blogmore.markdown.external_links import ExternalLinksExtension
 from blogmore.markdown.heading_anchors import HeadingAnchorsExtension
+from blogmore.markdown.mermaid import MermaidExtension
 from blogmore.markdown.optimised_images import OptimisedImagesExtension
 from blogmore.markdown.strikethrough import StrikethroughExtension
 
@@ -44,6 +45,7 @@ def create_custom_extensions(
         ExternalLinksExtension(site_url=site_url),
         HeadingAnchorsExtension(),
         StrikethroughExtension(),
+        MermaidExtension(),
     ]
     if with_optimised_images and image_manager is not None:
         extensions.append(